requirements.txt
cffi==1.16.0
EasyProcess==1.1
entrypoint2==1.1
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.1
Pillow==10.1.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pystray==0.19.5
pytweening==1.0.7
pywin32==306
six==1.16.0
sounddevice==0.4.6
tk==0.1.0
win10toast==0.9
